+++ src/payload.config.ts
@@ -5,6 +5,7 @@ import { Users } from './collections/Users'
 import { Categories } from './collections/Categories'
 import { Technologies } from './collections/Technologies'
 import { Sources } from './collections/Sources'
+import { Relations } from './collections/Relations'
 import { readServerEnvironment } from './lib/env'
 
 const environment = readServerEnvironment()
@@ -13,7 +14,7 @@ export default buildConfig({
   admin: {
     user: Users.slug,
   },
-  collections: [Users, Categories, Sources, Technologies],
+  collections: [Users, Categories, Sources, Technologies, Relations],
   db: postgresAdapter({
     idType: 'uuid',
     push: process.env.NODE_ENV === 'test',